Cleared useless static_cast; Changed EN_LOGS -> BOOST_LOGS; from_float -> from_arithmetic_t; Changed some types

+++ include/clib/VideoView.hpp
@@ -33,7 +33,7 @@ protected:
 
 public:
     using pixel_t = int;
-    using idx_t = unsigned;
+    using idx_t = size_t;
 
     // colors number
     enum spectrum
